+++ src/components/graph/BarChart.tsx
