+++ src/components/graph/DetailHeader.tsx
